Revert "[fix] 구매하기,구매내역페이지 : PurInfoBox 데이터 충돌로 인한 분리 및 코드 수정"
